Made constants file mutacc_auto/files/constants.py Support 'conda activate' for environments
